Add files via upload
Terminal -	It is an interface that allows users to interact with the system. In modern systems it is usually a terminal emulator window (like GNOME Terminal, Konsole, Windows Terminal).
Shell -	Program that actually interprets your commands. Example: Bash, Zsh, Fish. It runs inside the terminal.
Console - Physical device or system console. Traditionally refers to the text-based screen directly connected to the machine.
Prompt -	The text that the shell displays to indicate it’s ready to accept a command. Example: mayuri@ubuntu:~$ is a shell prompt.